+++ src/main/java/dev/ultimatchamp/bettergrass/mixin/ModelLoaderMixin.java
@@ -1,4 +1,5 @@
-package dev.ultimatchamp.bettergrass.mixin;
+//? if <1.21.2 {
+/*package dev.ultimatchamp.bettergrass.mixin;
 
 import dev.ultimatchamp.bettergrass.model.BetterGrassifyUnbakedModel;
 import dev.ultimatchamp.bettergrass.config.BetterGrassifyConfig;
@@ -15,76 +16,73 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import java.util.Map;
 
 //? if <1.21 {
-/*import net.minecraft.util.Identifier;
+/^import net.minecraft.util.Identifier;
 import java.util.Set;
-*///?}
+^///?}
 
 @Mixin(ModelLoader.class)
-public class BetterGrassifyModelLoaderMixin {
+public class ModelLoaderMixin {
     @Shadow
     @Final
     //? if >1.20.6 {
     private Map<ModelIdentifier, UnbakedModel> modelsToBake;
     //?} else {
-    /*private Map<Identifier, UnbakedModel> unbakedModels;
+    /^private Map<Identifier, UnbakedModel> unbakedModels;
 
     @Shadow
     @Final
     private Set<Identifier> modelsToLoad;
-    *///?}
+    ^///?}
 
     //? if >1.20.6 {
     @Inject(method = "addModelToBake", at = @At("HEAD"), cancellable = true)
-    private void onAddModelToBake(ModelIdentifier id, UnbakedModel unbakedModel, CallbackInfo ci) {
+    private void bettergrass$onAddModelToBake(ModelIdentifier id, UnbakedModel unbakedModel, CallbackInfo ci) {
     //?} else {
-    /*@Inject(method = "putModel", at = @At("HEAD"), cancellable = true)
-    private void onPutModel(Identifier id, UnbakedModel unbakedModel, CallbackInfo ci) {
-    *///?}
+    /^@Inject(method = "putModel", at = @At("HEAD"), cancellable = true)
+    private void bettergrass$onPutModel(Identifier id, UnbakedModel unbakedModel, CallbackInfo ci) {
+    ^///?}
         if (id instanceof ModelIdentifier modelId) {
             if (!modelId.getVariant().equals("inventory")) {
-                BetterGrassifyConfig.instance().moreBlocks.forEach(s -> {
-                    if (modelId.toString().startsWith(s.split("\\[")[0]) && !modelId.toString().contains("snowy=true")) {
+                for (String path : BetterGrassifyConfig.instance().moreBlocks) {
+                    if (modelId.toString().startsWith(path) && !modelId.toString().contains("snowy=true")) {
                         var newModel = new BetterGrassifyUnbakedModel(unbakedModel);
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
+                        ci.cancel();
+                    } if (modelId.toString().startsWith(path) && modelId.toString().contains("snowy=true") && BetterGrassifyConfig.instance().snowy) {
+                        var newModel = new BetterGrassifyUnbakedModel(unbakedModel);
+                        //? if >1.20.6 {
+                        this.modelsToBake.put(id, newModel);
+                        //?} else {
+                        /^this.unbakedModels.put(id, newModel);
+                        this.modelsToLoad.addAll(newModel.getModelDependencies());
+                        ^///?}
                         ci.cancel();
                     }
-                    if (BetterGrassifyConfig.instance().snowy) {
-                        if (modelId.toString().startsWith(s.split("\\[")[0]) && modelId.toString().contains("snowy=true")) {
-                            var newModel = new BetterGrassifyUnbakedModel(unbakedModel);
-                            //? if >1.20.6 {
-                            this.modelsToBake.put(id, newModel);
-                            //?} else {
-                            /*this.unbakedModels.put(id, newModel);
-                            this.modelsToLoad.addAll(newModel.getModelDependencies());
-                            *///?}
-                            ci.cancel();
-                        }
-                    }
-                });
+                }
 
                 if (BetterGrassifyConfig.instance().grassBlocks) {
-                    if (modelId.toString().startsWith("minecraft:grass_block".split("\\[")[0]) && !modelId.toString().contains("snowy=true")) {
+                    if (modelId.toString().startsWith("minecraft:grass_block") && !modelId.toString().contains("snowy=true")) {
                         var newModel = new BetterGrassifyUnbakedModel(unbakedModel);
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     } else if (modelId.toString().startsWith("minecraft:grass_block") && modelId.toString().contains("snowy=true") && BetterGrassifyConfig.instance().snowy) {
                         var newModel = new BetterGrassifyUnbakedModel(unbakedModel);
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     }
                 }
@@ -95,9 +93,9 @@ public class BetterGrassifyModelLoaderMixin {
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     }
                 }
@@ -108,9 +106,9 @@ public class BetterGrassifyModelLoaderMixin {
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     }
                 }
@@ -121,18 +119,18 @@ public class BetterGrassifyModelLoaderMixin {
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     } else if (modelId.toString().startsWith("minecraft:podzol") && modelId.toString().contains("snowy=true") && BetterGrassifyConfig.instance().snowy) {
                         var newModel = new BetterGrassifyUnbakedModel(unbakedModel);
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     }
                 }
@@ -143,18 +141,18 @@ public class BetterGrassifyModelLoaderMixin {
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     } else if (modelId.toString().startsWith("minecraft:mycelium") && modelId.toString().contains("snowy=true") && BetterGrassifyConfig.instance().snowy) {
                         var newModel = new BetterGrassifyUnbakedModel(unbakedModel);
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     }
                 }
@@ -165,9 +163,9 @@ public class BetterGrassifyModelLoaderMixin {
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     }
                 }
@@ -178,9 +176,9 @@ public class BetterGrassifyModelLoaderMixin {
                         //? if >1.20.6 {
                         this.modelsToBake.put(id, newModel);
                         //?} else {
-                        /*this.unbakedModels.put(id, newModel);
+                        /^this.unbakedModels.put(id, newModel);
                         this.modelsToLoad.addAll(newModel.getModelDependencies());
-                        *///?}
+                        ^///?}
                         ci.cancel();
                     }
                 }
@@ -188,3 +186,4 @@ public class BetterGrassifyModelLoaderMixin {
         }
     }
 }
+*///?}